[refactor] 예외 처리 도메인으로 위치 이동
