+++ src/main/java/lotto/domain/InvalidNumberException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain;
 
 public class InvalidNumberException extends RuntimeException{
     private static final String INVALID_NUMBER = "숫자가 잘못되었습니다";
